renaming features to WhatWeBring
