+++ components/WhatWeBring/index.tsx
@@ -1,10 +1,10 @@
 "use client";
 import React from "react";
-import featuresData from "./featuresData";
-import SingleFeature from "./SingleFeature";
+import PanelData from "./PanelData";
+import SingleFeature from "./SinglePanel";
 import SectionHeader from "../Common/SectionHeader";
 
-const Feature = () => {
+const WhatWeBring = () => {
   return (
     <>
       {/* <!-- ===== Features Start ===== --> */}
@@ -23,7 +23,7 @@ const Feature = () => {
           <div className="mt-12.5 grid grid-cols-1 gap-7.5 md:grid-cols-2 lg:mt-15 lg:grid-cols-3 xl:mt-20 xl:gap-12.5">
             {/* <!-- Features item Start --> */}
 
-            {featuresData.map((feature, key) => (
+            {PanelData.map((feature, key) => (
               <SingleFeature feature={feature} key={key} />
             ))}
             {/* <!-- Features item End --> */}
@@ -36,4 +36,4 @@ const Feature = () => {
   );
 };
 
-export default Feature;
+export default WhatWeBring;
